+++ Source/Scene/LabelCollection.js
@@ -76,6 +76,8 @@ define([
         if (defined(billboard)) {
             billboard.setShow(false);
             billboard.setImageIndex(-1);
+            // Destroy pickId to allow setting _pickIdThis and _id when the billboard is reused.
+            billboard._pickId = billboard._pickId && billboard._pickId.destroy();
             labelCollection._spareBillboards.push(billboard);
             glyph.billboard = undefined;
         }
@@ -178,6 +180,7 @@ define([
                     billboard.setVerticalOrigin(label._verticalOrigin);
                     billboard.setScale(label._scale);
                     billboard._pickIdThis = label;
+                    billboard._id = label._id;
                 }
 
                 glyph.billboard.setImageIndex(glyphTextureInfo.index);