+++ Specs/Scene/LabelCollectionSpec.js
@@ -85,6 +85,7 @@ defineSuite([
         expect(label.getHorizontalOrigin()).toEqual(HorizontalOrigin.LEFT);
         expect(label.getVerticalOrigin()).toEqual(VerticalOrigin.BOTTOM);
         expect(label.getScale()).toEqual(1.0);
+        expect(label.getId()).not.toBeDefined();
         expect(label.getTranslucencyByDistance()).not.toBeDefined();
     });
 
@@ -128,6 +129,7 @@ defineSuite([
             horizontalOrigin : horizontalOrigin,
             verticalOrigin : verticalOrigin,
             scale : scale,
+            id : 'id',
             translucencyByDistance : translucency
         });
 
@@ -144,6 +146,7 @@ defineSuite([
         expect(label.getHorizontalOrigin()).toEqual(horizontalOrigin);
         expect(label.getVerticalOrigin()).toEqual(verticalOrigin);
         expect(label.getScale()).toEqual(scale);
+        expect(label.getId()).toEqual('id');
         expect(label.getTranslucencyByDistance()).toEqual(translucency);
     });
 
@@ -682,11 +685,13 @@ defineSuite([
             },
             text : 'x',
             horizontalOrigin : HorizontalOrigin.CENTER,
-            verticalOrigin : VerticalOrigin.CENTER
+            verticalOrigin : VerticalOrigin.CENTER,
+            id : 'id'
         });
 
         var pickedObject = pick(context, frameState, labels, 0, 0);
         expect(pickedObject.primitive).toEqual(label);
+        expect(pickedObject.id).toEqual('id');
     });
 
     it('does not pick a label with show set to false', function() {